refactor: split reference collection into extract/references.rs
extract.rs (1149 lines) and extract_tests/rust.rs (1009 lines) both
crossed the warn threshold after ADR 0064; reference-name collection
is an independent responsibility, so it moves out along that boundary
(ADR 0028) together with its tests.
